Add pixel-precise drop-in implementation of the TFT_eSPI display interface to allow for rendering thumbnails generated by Lopaka. The current minimal functionality allows for rendering of a simple sudoku grid.

+++ src/platform/emulator/sfml_display.hpp
@@ -131,8 +131,8 @@ class SfmlDisplay : public Display, public TftCompatibleDisplay
                       uint32_t bg, uint8_t size) override;
         void drawLine(int32_t xs, int32_t ys, int32_t xe, int32_t ye,
                       uint32_t color) override;
-        void drawRect(int32_t x, int32_t y, int32_t w, int32_t h,
-                      uint32_t color) override;
+        void drawRect(int x, int y, int w, int h,
+                      int color) override;
         void fillRect(int32_t x, int32_t y, int32_t w, int32_t h,
                       uint32_t color) override;
         void drawRoundRect(int32_t x, int32_t y, int32_t w, int32_t h,
@@ -145,7 +145,7 @@ class SfmlDisplay : public Display, public TftCompatibleDisplay
                         uint32_t color) override;
         void drawString(const char *string, int32_t x, int32_t y) override;
         void fillScreen(uint32_t color) override;
-        void setTextColor(uint16_t color) override;
+        void setTextColor(uint32_t color) override;
         void setTextSize(uint8_t size) override;
 
       private:
